ticketsapi/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.0.12
djangorestframework==3.10.2

djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.8
pipreqs==0.4.9
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
urllib3==1.25.3
yarg==0.1.9
